pyproject.toml
[tool.poetry]
name = "autocsr"
version = "1.0.9"
description = "AutoCSR is a command-line tool and library for automatically generating Certificate Signing Requests from easy to define configuration files."
authors = ["Max Wolfe <max@securitywolfe.com>"]
readme = "README.md"
license = "MIT"
repository = "https://github.com/maxwolfe/autocsr"
documentation = "https://autocsr.readthedocs.io/en/latest/"
keywords = [
  "Certificate Signing Request",
  "CSR",
]

[tool.poetry.dependencies]
python = "^3.7"
cryptography = "^3.4.7"
protobuf = "^3.15.8"
PyYAML = "^5.4.1"
typer = "^0.3.2"
Jinja2 = "^2.11.3"
pyOpenSSL = "^20.0.1"
pyasn1-modules = "^0.2.8"
python-pkcs11 = "^0.7.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
yamllint = "^1.26.1"
black = "^20.8b1"
isort = "^5.8.0"
python-language-server = {extras = ["all"], version = "^0.36.2"}
flake8 = "3.8.4"
coverage = "^5.5"
Sphinx  = "^3.5.4"
sphinx-rtd-theme = "^0.5.2"
autodoc = "^0.5.0"

[tool.poetry.scripts]
autocsr = "autocsr.cli:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.urls]
issues = "https://github.com/maxwolfe/autocsr/issues"
